breakout massive services into shareable mixins and classes, fix training
