+++ app/components/user-training-item/component.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
 
 export default Ember.Component.extend({
-  classNames: ['panel', 'panel-default', 'user', 'resource-list-item'],
+  classNames: ['panel', 'panel-default', 'user', 'resource-list-item']
 });
