+++ app/utils/generic-criterion-status.js
@@ -1,0 +1,6 @@
+import Ember from 'ember';
+import CriterionStatusMixin from 'diesel/mixins/services/criterion-status';
+
+export default Ember.Object.extend(CriterionStatusMixin, {
+  showDefaultStatus: true
+});
