+++ tests/unit/utils/training-enrollment-status-test.js
@@ -1,6 +1,5 @@
 import { test } from 'ember-qunit';
 import Ember from 'ember';
-import { fromNow, ago } from '../../helpers/date';
 import TrainingEnrollmentStatus from 'diesel/utils/training-enrollment-status';
 
 module('Unit: TrainingEnrollmentStatus');
